+++ bridge/opentracing/util.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package opentracing
+package opentracing // import "go.opentelemetry.io/otel/bridge/opentracing"
 
 import (
 	"context"