+++ propagators/trace_context_benchmark_test.go
@@ -38,14 +38,10 @@ func BenchmarkInject(b *testing.B) {
 
 func injectSubBenchmarks(b *testing.B, fn func(context.Context, *testing.B)) {
 	b.Run("SampledSpanContext", func(b *testing.B) {
-		var id uint64
 		spanID, _ := otel.SpanIDFromHex("00f067aa0ba902b7")
 		traceID, _ := otel.TraceIDFromHex("4bf92f3577b34da6a3ce929d0e0e4736")
 
-		mockTracer := &oteltest.MockTracer{
-			Sampled:     false,
-			StartSpanID: &id,
-		}
+		mockTracer := oteltest.DefaultTracer()
 		b.ReportAllocs()
 		sc := otel.SpanContext{
 			TraceID:    traceID,