+++ sdk/instrumentation/library.go
@@ -16,10 +16,14 @@
 Package instrumentation provides an instrumentation library structure to be
 passed to both the OpenTelemetry Tracer and Meter components.
 
+This package is currently in a pre-GA phase. Backwards incompatible changes
+may be introduced in subsequent minor version releases as we work to track the
+evolving OpenTelemetry specification and user feedback.
+
 For more information see
 [this](https://github.com/open-telemetry/oteps/blob/master/text/0083-component.md).
 */
-package instrumentation
+package instrumentation // import "go.opentelemetry.io/otel/sdk/instrumentation"
 
 // Library represents the instrumentation library.
 type Library struct {