+++ sdk/metric/controller/push/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package push
+package push // import "go.opentelemetry.io/otel/sdk/metric/controller/push"
 
 import (
 	"time"