+++ sdk/metric/controller/push/push.go
@@ -61,7 +61,7 @@ func New(checkpointer export.Checkpointer, exporter export.Exporter, opts ...Opt
 
 	impl := sdk.NewAccumulator(
 		checkpointer,
-		sdk.WithResource(c.Resource),
+		c.Resource,
 		sdk.WithMetricsProcessors(c.MetricsProcessors...),
 	)
 	return &Controller{